RecurrenceMicrostatesAnalysis.jl

RecurrenceMicrostatesAnalysis — Module

RecurrenceMicrostatesAnalysis.jl

RecurrenceMicrostatesAnalysis.jl is a simple and fast Julia-based package for recurrence microstates analysis. It implements the computation of Recurrence Microstates Analysis (RMA) distributions, specific quantifiers — such as disorder — and the estimation of typical RQA quantifiers, including determinism and laminarity.

RMA is a subfield of Recurrence Analysis and is a powerful tool for analyzing large time series or large datasets using statistical methods, offering high performance and avoiding memory issues. Although the field is still relatively new, it has shown promising applications, including in Machine Learning.

RecurrenceMicrostatesAnalysis.jl is part of DynamicalSystems.jl, and can be used either as a standalone package or in combination with other packages in the ecosystem. It is also integrated with ComplexityMeasures.jl, providing an easy-to-use interface.

To install the package, run:

import Pkg
Pkg.add("RecurrenceMicrostatesAnalysis")

Citation

If you use this package in a publication, or simply want to refer to it, please cite the paper below:

@article{RecurrenceMicrostatesAnalysis.jl,
    author = {Vinicius Ferreira, Gabriel and Lopes da Cruz, Felipe Eduardo and Marghoti, Gabriel and de Lima Prado, Thiago and Roberto Lopes, Sergio and Marwan, Norbert and Kurths, Jürgen},
    title = {RecurrenceMicrostatesAnalysis.jl: A Julia library for analyzing dynamical systems with recurrence microstates},
    journal = {Chaos: An Interdisciplinary Journal of Nonlinear Science},
    volume = {35},
    number = {11},
    pages = {113123},
    year = {2025},
    month = {11},
    issn = {1054-1500},
    doi = {10.1063/5.0293708},
    url = {https://doi.org/10.1063/5.0293708}
}

Latest news

RecurrenceMicrostateAnalysis.jl is now integrated into DynamicalSystems.jl.

Realizing that RMA are complexity measures, the API of RecurrenceMicrostateAnalysis.jl has been updated to inherit and extend from ComplexityMeasures.jl.

Brand new comprehensive tutorial.

You can always see the CHANGELOG.md (at the GitHub repo) for more!

Getting started

Start by reviewing the tutorial for a basic introduction to using the package. You can also consult individual functions in the API, and find applied examples in the dedicated Examples section.

We also provide a section RecurrenceMicrostatesAnalysis.jl for devs for those interested in developing new methods for RecurrenceMicrostatesAnalysis.jl, such as new microstate shapes, sampling modes, recurrence functions, or complexity estimators.